backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-socketio==5.11.2
python-dotenv==1.0.1
numpy==1.26.4
openpyxl==3.1.2